requirements.txt
Django
djangorestframework
celery
redis
psycopg2-binary>=2.8.5
beautifulsoup4
lxml
requests
requests-html
pytest
pytest-asyncio
pytest-django
django-cors-headers
celery-progress